separate file for PaymentMethod class
