+++ lib/src/responses.dart
@@ -1,3 +1,4 @@
+import './responses/payment_method.dart';
 part './responses/card.dart';
 part './responses/token.dart';
 part './responses/source.dart';
